Fix regression with system package manager tools (#18872)
* package_manager (Apt): Fix check_command for all packages

The current `check_command` for Apt at foreign architectures checks
for `<package_name:<arch>`.
However, this selector doesn't find all packages.

The solution is to just check for the package name,
dpkg-query will list the architecures and we check for the correct one.

Docs: omit
Close #18569

* package_manager (Apt): Fix check_command for multiple matches

When `dpkg-query` finds multiple matches, it printed that on the same line,
which broke the grep.

Close #18862

* pass base_name

* fix tests

* Conan 2.20.1

* Update conan/tools/system/package_manager.py

* fix potential issue with \n using raw strings

---------

Co-authored-by: Alexander Krabler <alexander.krabler@kuka.com>