Fix arch used Apt.check with host_pacakge=False, when cross-compiling (#19074)
* Add Apt.check tests with host_package=False

* Fix Apt dpkg query arch when host_package=False

Apt.check(<pkg>, host_package=False) added the host arch,
instead of the build arch to the dpkg query, if used in
a recipe during cross-building conan install invocations.
This fix ensures that the `settings_build.arch` is used
to make the query in this context.

Issue: https://github.com/conan-io/conan/issues/19073

* Add missing arch setting in test_tools_check

The mocked settings was missing an `arch`,
causing `_SystemPackageManagerTool._arch` to
be unset in the tests, leading to erroneous
test failures.

* Apply suggestion from @czoido

---------

Co-authored-by: Carlos Zoido <mrgalleta@gmail.com>